+++ example/lib/styles/styles.dart
@@ -72,7 +72,7 @@ class _ChatPageState extends State<ChatPage>
     );
 
     final halloweenActionButtonStyle = ActionButtonStyle(
-      tooltipTextStyle: halloweenTextStyle,
+      textStyle: halloweenTextStyle,
       iconColor: Colors.black,
       iconDecoration: BoxDecoration(
         color: Colors.orange,
@@ -81,7 +81,7 @@ class _ChatPageState extends State<ChatPage>
     );
 
     final halloweenMenuButtonStyle = ActionButtonStyle(
-      tooltipTextStyle: halloweenTextStyle,
+      textStyle: halloweenTextStyle,
       iconColor: Colors.orange,
       iconDecoration: BoxDecoration(
         color: Colors.black,