+++ lib/src/views/chat_input/editing_indicator.dart
@@ -3,7 +3,7 @@ import 'package:flutter/widgets.dart';
 import '../../styles/action_button_style.dart';
 import '../../styles/toolkit_text_styles.dart';
 import '../../utility.dart';
-import '../action_button/action_button.dart';
+import '../action_button.dart';
 
 /// A widget that displays an editing indicator with a cancel button.
 ///